refactor: #7 BaseEntity 재정의
