--- src/main/java/com/goti/domain/base/BaseEntity.java
@@ -1,45 +1,0 @@
-package com.goti.domain.base;
-
-import java.time.Instant;
-import java.util.UUID;
-
-import org.hibernate.annotations.UuidGenerator;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.domain.Persistable;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.Id;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity implements Persistable<UUID> {
-
-	@Id
-	@UuidGenerator
-	@Column(name = "id", nullable = false, updatable = false)
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	private UUID id;
-
-	@CreatedDate
-	@Column(nullable = false, updatable = false)
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	private Instant createdAt;
-
-	@LastModifiedDate
-	@Column(nullable = false)
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	private Instant updatedAt;
-
-	@Override
-	public boolean isNew() {
-		return id == null;
-	}
-}
