+++ src/main/java/com/goti/domain/base/CreationTimestampEntity.java
@@ -1,0 +1,23 @@
+package com.goti.domain.base;
+
+import java.time.Instant;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class CreationTimestampEntity extends BaseUuidEntity {
+
+	@CreatedDate
+	@Column(nullable = false, updatable = false)
+	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
+	private @Getter Instant createdAt;
+}
